clean up peer utils
