--- peers/utils.go
@@ -1,68 +1,0 @@
-package peers
-
-import (
-	"github.com/ava-labs/avalanchego/ids"
-	"github.com/ava-labs/avalanchego/utils/set"
-	"github.com/ava-labs/avalanchego/vms/platformvm/warp"
-	"github.com/ava-labs/awm-relayer/validators"
-)
-
-// Helper struct to hold connected validator information
-// Warp Validators sharing the same BLS key may consist of multiple nodes,
-// so we need to track the node ID to validator index mapping
-type ConnectedCanonicalValidators struct {
-	ConnectedWeight       uint64
-	TotalValidatorWeight  uint64
-	ValidatorSet          []*warp.Validator
-	nodeValidatorIndexMap map[ids.NodeID]int
-}
-
-// Returns the Warp Validator and its index in the canonical Validator ordering for a given nodeID
-func (c *ConnectedCanonicalValidators) GetValidator(nodeID ids.NodeID) (*warp.Validator, int) {
-	return c.ValidatorSet[c.nodeValidatorIndexMap[nodeID]], c.nodeValidatorIndexMap[nodeID]
-}
-
-// ConnectToCanonicalValidators connects to the canonical validators of the given subnet and returns the connected
-// validator information
-func ConnectToCanonicalValidators(
-	network *AppRequestNetwork,
-	client *validators.CanonicalValidatorClient,
-	subnetID ids.ID,
-) (*ConnectedCanonicalValidators, error) {
-	// Get the subnet's current canonical validator set
-	validatorSet, totalValidatorWeight, err := client.GetCurrentCanonicalValidatorSet(subnetID)
-	if err != nil {
-		return nil, err
-	}
-
-	// We make queries to node IDs, not unique validators as represented by a BLS pubkey, so we need this map to track
-	// responses from nodes and populate the signatureMap with the corresponding validator signature
-	// This maps node IDs to the index in the canonical validator set
-	nodeValidatorIndexMap := make(map[ids.NodeID]int)
-	for i, vdr := range validatorSet {
-		for _, node := range vdr.NodeIDs {
-			nodeValidatorIndexMap[node] = i
-		}
-	}
-
-	// Manually connect to all peers in the validator set
-	// If new peers are connected, AppRequests may fail while the handshake is in progress.
-	// In that case, AppRequests to those nodes will be retried in the next iteration of the retry loop.
-	nodeIDs := set.NewSet[ids.NodeID](len(nodeValidatorIndexMap))
-	for node := range nodeValidatorIndexMap {
-		nodeIDs.Add(node)
-	}
-	connectedNodes := network.ConnectPeers(nodeIDs)
-
-	// Check if we've connected to a stake threshold of nodes
-	connectedWeight := uint64(0)
-	for node := range connectedNodes {
-		connectedWeight += validatorSet[nodeValidatorIndexMap[node]].Weight
-	}
-	return &ConnectedCanonicalValidators{
-		ConnectedWeight:       connectedWeight,
-		TotalValidatorWeight:  totalValidatorWeight,
-		ValidatorSet:          validatorSet,
-		nodeValidatorIndexMap: nodeValidatorIndexMap,
-	}, nil
-}
